refactor: move bank check transaction logic to new module
